Moved Activity to its own folder
