+++ components/Activity/Activity.jsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { Text } from "react-native"
 
-import Layout from "./shared/Layout.jsx"
+import Layout from "../shared/Layout.jsx"
 
 const Activity = () => {
     return (
